Inject ContractMarketInfo to each member tab, show placeholder for pages that are sweden only (so far)
