+++ src/components/member/tabs/index.js
@@ -1,4 +1,3 @@
-import { AccountTab } from 'components/member/tabs/AccountTab'
 import ChatPane from 'components/member/tabs/ChatPane'
 import ClaimsTab from 'components/member/tabs/ClaimsTab'
 import { MemberDebtComponent } from 'components/member/tabs/DebtTab'
@@ -15,6 +14,8 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { Tab } from 'semantic-ui-react'
 import styled from 'react-emotion'
+import { useContractMarketInfo } from 'graphql/use-get-member-contract-market-info'
+import { AccountTab } from './account-tab'
 
 const TabContainer = styled(Tab.Pane)`
   &&& {
@@ -25,11 +26,15 @@ const TabContainer = styled(Tab.Pane)`
   }
 `
 
-const TabItem = ({ props, TabContent }) => (
-  <TabContainer>
-    <TabContent {...props} />
-  </TabContainer>
-)
+const TabItem = ({ props, TabContent }) => {
+  const memberId = props.match.params.memberId
+  const [contractMarketInfo] = useContractMarketInfo(memberId)
+  return (
+    <TabContainer>
+      <TabContent {...props} contractMarketInfo={contractMarketInfo}/>
+    </TabContainer>
+  )
+}
 
 TabItem.propTypes = {
   props: PropTypes.object.isRequired,
@@ -38,32 +43,35 @@ TabItem.propTypes = {
   hideTab: PropTypes.bool,
 }
 
-/* eslint-disable react/display-name */
 const memberPagePanes = (props, addMessage, socket) => {
   const { insurance } = props
+  const memberId = props.match.params.memberId
   const panes = [
     {
-      menuItem: 'Details',
-      render: () => <TabItem props={props} TabContent={DetailsTab} />,
+      menuItem: 'Member',
+      render: () => (
+        <TabItem props={props} TabContent={DetailsTab} />
+      ),
     },
     {
       menuItem: 'Claims',
-      render: () => <TabItem props={props} TabContent={ClaimsTab} />,
+      render: () => (
+        <TabItem props={props} TabContent={ClaimsTab} />
+      ),
     },
     {
       menuItem: 'Files',
-      render: () => <TabItem props={props} TabContent={MemberFile} />,
-    },
-    {
-      menuItem: 'Debt',
-      render: () => <TabItem props={props} TabContent={MemberDebtComponent} />,
+      render: () => (
+        <TabItem props={props} TabContent={MemberFile} />
+      ),
     },
     {
       menuItem: 'Tickets',
       render: () => (
         <TabItem
           props={{
-            memberId: props.match.params.memberId,
+            ...props,
+            memberId,
             ticketType: 'REMIND',
           }}
           TabContent={CreateTicketStandAlone}
@@ -88,7 +96,7 @@ const memberPagePanes = (props, addMessage, socket) => {
       menuItem: 'Contracts',
       render: () => (
         <TabItem
-          props={{ memberId: props.match.params.memberId }}
+          props={{ ...props, memberId }}
           TabContent={Contracts}
         />
       ),
@@ -96,10 +104,7 @@ const memberPagePanes = (props, addMessage, socket) => {
     {
       menuItem: 'Quotes',
       render: () => (
-        <TabItem
-          props={{ memberId: props.match.params.memberId }}
-          TabContent={Quotes}
-        />
+        <TabItem props={{ ...props, memberId }} TabContent={Quotes} />
       ),
     },
   )
@@ -107,22 +112,47 @@ const memberPagePanes = (props, addMessage, socket) => {
     panes.push(
       {
         menuItem: 'Current Insurance',
-        render: () => <TabItem props={props} TabContent={InsuranceTab} />,
+        render: () => (
+          <TabItem
+            props={props}
+            TabContent={InsuranceTab}
+          />
+        ),
       },
       {
         menuItem: 'All Insurances',
-        render: () => <TabItem props={props} TabContent={InsuranceListTab} />,
-      },
-      {
-        menuItem: 'Payments',
-        render: () => <TabItem props={props} TabContent={PaymentsTab} />,
-      },
-      {
-        menuItem: 'Account',
-        render: () => <TabItem props={props} TabContent={AccountTab} />,
+        render: () => (
+          <TabItem
+            props={props}
+            TabContent={InsuranceListTab}
+          />
+        ),
       },
     )
   }
+  panes.push(
+    {
+      menuItem: 'Payments',
+      render: () => (
+        <TabItem memberId={memberId} props={props} TabContent={PaymentsTab} />
+      ),
+    },
+    {
+      menuItem: 'Account',
+      render: () => (
+        <TabItem memberId={memberId} props={{...props, memberId}} TabContent={AccountTab} />
+      ),
+    },
+    {
+      menuItem: 'Debt',
+      render: () => (
+        <TabItem
+          props={props}
+          TabContent={MemberDebtComponent}
+        />
+      ),
+    },
+  )
 
   return panes
 }
